+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 0.2.3 (Next)
+### 0.3.0 (Next)
 
 * [#70](https://github.com/ruby-grape/grape-swagger-rails/pull/70): Rails 5 support - [@serggl](https://github.com/serggl).
 * [#68](https://github.com/ruby-grape/grape-swagger-rails/pull/68): Added danger, PR linter - [@dblock](https://github.com/dblock).